Add logic for implementing HO100212 exception
The test for HO100212 is currently skipped, as it will fail when
running against Bichard. This is because if the person's title is too
long, which is what the HO100212 exception is for, Bichard doesn't
actually throw the exception but instead fails to parse the XML and
completely gives up.