Add inline npmrc file with registry config during GitHub Action
Despite what the docs [1] say, it seems that this file is still required
for publishing to npm. However, the presence of this file causes
codebuild to complain about the `NPM_TOKEN` variable not being set.

Therefore, this commit generates that file only when needed, and doesn't
commit it to the repository.

[1] https://docs.npmjs.com/using-private-packages-in-a-ci-cd-workflow