Refactor resubmit endpoint and related functions
This commit refactors the resubmit endpoint and related functions in the `cases` route. It introduces a new function `canUserResubmitCase` to check if a user is allowed to resubmit a case. The `formatForceNumbers` function is also modified to handle `null` and `undefined` values. Unit tests are added for both the `formatForceNumbers` and `canUserResubmitCase` functions.