Simplify convertHoOffenceCodeToPncFormat
The PNC format for an offence code is the offence code for a local offence
and a concatenation of properties for a national offence reason which is
actually just the `FullCode`.

This is actually built up when we parse the AHO meaning we don't need to do
it again here.