Refactor hearing details functions (#1196)
* Move hearing detail functions into their own files

* Add tests for preProcessOffenceReasonSequence

* Remove unnecessary empty string default in createArrestHearingFromOffence

* Simplify convertHoOffenceCodeToPncFormat

The PNC format for an offence code is the offence code for a local offence
and a concatenation of properties for a national offence reason which is
actually just the `FullCode`.

This is actually built up when we parse the AHO meaning we don't need to do
it again here.

* Add tests for convertHoOffenceCodeToPncFormat

* Add tests for getOffenceCode

* Refactor getOffenceCode

* Remove convertHoOffenceCodeToPncFormat and use getOffenceCode instead

Turns out we already have an existing function that does the same thing.

* Move getOffenceCode to root lib folder

* Refactor to use getOffenceCode from Core in UI

And remove duplicated function.

* Add tests for createCourtHearingFromOffence

* Add tests for createAdjudicationFromOffence

* Remove unnecessary nullish coalescing assignment in createAdjudicationFromOffence

* Move createPncDisposalFromOffence into createDisposalsFromOffence

As createPncDisposalFromOffence is only used for createDisposalsFromOffence.

* Remove unused hoOffencesAreEqual function