Fix sidebar tab panels not being keyboard accessible
When the accessibility audit was conducted, using the tab key on the
case details page through the court case details tabs to the sidebar,
it would focus straight to an element within the current sidebar tab.

This has since been fixed when we removed govuk-react however you weren't
able to use the arrow keys to move between the sidebar tabs which this PR
fixes.

The GOVUK Frontend provides the ability to add `data-module="govuk-tabs"`
which handles adding the appropriate ARIA labels and role attributes that
the accessibility audit report mentions and allows arrow key switching.

However we can't use this as we aren't able to customise the default tab
we show when you go onto the case details page without putting focus
on a tab on the sidebar upon arrival.

As a result, we have to handle this manually.